Add SBOM action to build/release process.
This has been requested by customers to enable better security practices.

When run in a build job, it uploads the generated SBOM as a build artifact `opentelemetry-java-build.spdx.json`.
When run as a release, it will also upload the SBOM to the github release.

I tested the build part on a different repo, but didn't verify the release part.